actualizacion ui

+++ sputnik-ui/src/app/guard/auth.guard.ts
@@ -8,11 +8,11 @@ export class AuthGuard implements CanActivate {
   constructor(public auth: AuthService, private router: Router) {}
 
   canActivate(): boolean {
-    /*if (!this.auth.estaAutenticado()) {
+    if (!this.auth.estaAutenticado()) {
       localStorage.clear();
       this.router.navigate(['/login']);
       return false;
-    }*/
+    }
     return true;
   }
 }
